+++ R/select.R
@@ -15,6 +15,8 @@ get.MSPE <- function(CTMM,MSPE="position")
   else { MSPE <- Inf }
   return(MSPE)
 }
+
+
 ########
 get.IC <- function(CTMM,IC="AICc")
 {
@@ -22,6 +24,37 @@ get.IC <- function(CTMM,IC="AICc")
   else { IC <- Inf }
   return(IC)
 }
+
+
+##################
+# function to simplify complexity of models
+simplify.ctmm <- function(M,par)
+{
+  if("eccentricity" %in% par)
+  {
+    M$isotropic <- TRUE
+    M$sigma <- covm(M$sigma,isotropic=TRUE,axes=M$axes)
+    if("COV" %in% names(M)) { M$COV <- rm.name(M$COV,"eccentricity") }
+  }
+
+  if("circle" %in% par) { M$circle <- FALSE }
+
+  if("range" %in% par)
+  {
+    M$tau[1] <- Inf
+    M$range <- FALSE
+  }
+
+  # autocorrelation timescales can't be distinguished
+  if("diff.tau" %in% par)
+  {
+    M$tau <- c(1,1)/mean(1/M$tau)
+    M$omega <- FALSE
+  }
+
+  return(M)
+}
+
 
 ###############
 # keep removing uncertain parameters until AIC stops improving
@@ -57,26 +90,6 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
     return(MLE)
   }
 
-  # function to simplify complexity of models
-  simplify <- function(M,par)
-  {
-    if("eccentricity" %in% par)
-    {
-      M$isotropic <- TRUE
-      M$sigma <- covm(M$sigma,isotropic=TRUE)
-    }
-
-    if("circle" %in% par) { M$circle <- FALSE }
-
-    if("range" %in% par)
-    {
-      M$tau[1] <- Inf
-      M$range <- FALSE
-    }
-
-    return(M)
-  }
-
   # consider a bunch of new models and update best model without duplication
   iterate <- function(DROP,REFINE=list())
   {
@@ -93,7 +106,7 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
     GUESS <- c(DROP,REFINE)
 
     # fit every model
-    if(trace && length(GUESS)) { message("* Fitting models ",paste(names(GUESS),collapse=", ")) }
+    if(trace && length(GUESS)) { message("* Fitting models ",paste(names(GUESS),collapse=", "),".") }
     #? should I run select here instead of fit ?
     GUESS <- lapply(GUESS,function(g){ctmm.fit(data,g,trace=trace2,...)})
 
@@ -124,10 +137,11 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
   FEATURES <- FEATURES[!(FEATURES=="area")]
   FEATURES <- FEATURES[!(FEATURES=="error")]
   FEATURES <- FEATURES[!grepl("tau",FEATURES)]
+  FEATURES <- FEATURES[!(FEATURES=="omega")]
 
   # start with the most basic "compatible" model
-  GUESS <- simplify(CTMM,FEATURES)
-  if(trace) { message("* Fitting model ",name.ctmm(GUESS)) }
+  GUESS <- simplify.ctmm(CTMM,FEATURES)
+  if(trace) { message("* Fitting model ",name.ctmm(GUESS),".") }
   TARGET <- CTMM
   CTMM <- ctmm.fit(data,GUESS,trace=trace2,...)
   MODELS <- list(CTMM)
@@ -171,11 +185,11 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
   {
     GUESS <- list()
     MLE <- get.mle()
-
     beta <- alpha.ctmm(CTMM,alpha)
+
     # consider if some timescales are actually zero
     CI <- confint.ctmm(CTMM,alpha=beta)
-    if(length(CTMM$tau)==2 && !is.na(IC))
+    if(length(CTMM$tau)==2 && !CTMM$omega && CTMM$tau[1]!=CTMM$tau[2] && !is.na(IC))
     {
       Q <- CI["tau velocity",1]
       if(is.nan(Q) || (Q<=0))
@@ -194,11 +208,39 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
       }
     }
 
+    # can autocorrelation timescales be distinguished?
+    if(length(CTMM$tau)==2 && (CTMM$tau[1]!=CTMM$tau[2] || CTMM$omega))
+    {
+      TEMP <- get.taus(CTMM,zeroes=TRUE)
+      nu <- TEMP$f.nu[2] # frequency/difference
+      J <- TEMP$J.nu.tau[2,] # Jacobian of nu WRT canonical parameters
+      Q <- TEMP$tau.names
+      Q <- c(J %*% CTMM$COV[Q,Q] %*% J) # variance of nu
+      Q <- ci.tau(nu,Q,alpha=beta)[1]
+
+      if(Q<=0 || is.na(IC))
+      { GUESS <- c(GUESS,list(simplify.ctmm(MLE,"diff.tau"))) }
+    }
+    else if(length(CTMM$tau)==2) # try other side if boundary if choosen model is critically damped
+    {
+      # try overdamped
+      TEMP <- MLE
+      TEMP$omega <- 0
+      TEMP$tau <- TEMP$tau * exp(c(1,-1)*sqrt(.Machine$double.eps))
+      GUESS <- c(GUESS,list(TEMP))
+
+      # try underdamped
+      TEMP <- MLE
+      TEMP$tau <- c(1,1)/mean(1/TEMP$tau)
+      TEMP$omega <- sqrt(.Machine$double.eps)
+      GUESS <- c(GUESS,list(TEMP))
+    }
+
     # consider if there is no circulation
     if(CTMM$circle)
     {
       Q <- CI["circle",3]
-      if(is.nan(Q) || (Q==Inf)) { GUESS <- c(GUESS,list(simplify(MLE,"circle"))) }
+      if(is.nan(Q) || (Q==Inf) || is.na(IC)) { GUESS <- c(GUESS,list(simplify.ctmm(MLE,"circle"))) }
     }
 
     # consider if eccentricity is zero
@@ -206,12 +248,12 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
     {
       Q <- "eccentricity"
       Q <- stats::qnorm(beta/2,mean=CTMM$sigma@par[Q],sd=sqrt(CTMM$COV[Q,Q]))
-      if(Q <= 0) { GUESS <- c(GUESS,list(simplify(MLE,"eccentricity"))) }
+      if(Q<=0 || is.na(IC)) { GUESS <- c(GUESS,list(simplify.ctmm(MLE,"eccentricity"))) }
     }
 
     # consider if we can relax range residence (non-likelihood comparison only)
     if(CTMM$range && is.na(IC))
-    { GUESS <- c(GUESS,list(simplify(MLE,"range"))) }
+    { GUESS <- c(GUESS,list(simplify.ctmm(MLE,"range"))) }
 
     # consider if the mean could be more detailed
     REFINE <- drift@refine(MLE)
@@ -224,6 +266,12 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
   if(verbose)
   {
     MODELS <- sort.ctmm(MODELS,IC=IC,MSPE=MSPE)
+
+    # remove redundant models
+    NAMES <- sapply(MODELS,name.ctmm)
+    KEEP <- c(TRUE, NAMES[-1]!=NAMES[-length(NAMES)] )
+    MODELS <- MODELS[KEEP]
+
     return(MODELS)
   }
   else
@@ -234,11 +282,17 @@ ctmm.select <- function(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",MSPE="posit
 name.ctmm <- function(CTMM,whole=TRUE)
 {
   # base model
-  if(length(CTMM$tau)==2)
-  { if(CTMM$tau[1]<Inf) { NAME <- "OUF" } else { NAME <- "IOU" } }
-  else if(length(CTMM$tau)==1)
-  { if(CTMM$tau[1]<Inf) { NAME <- "OU" } else { NAME <- "BM" } }
-  else if(length(CTMM$tau)==0)
+  tau <- CTMM$tau
+  if(length(tau)==2)
+  {
+    if(tau[1]==Inf) { NAME <- "IOU" }
+    else if(tau[1]>tau[2]) { NAME <- "OUF" }
+    else if(CTMM$omega) { NAME <- "OU\u03A9" } # underdamped
+    else { NAME <- "OUf" } # identical timescales
+  }
+  else if(length(tau)==1)
+  { if(tau[1]<Inf) { NAME <- "OU" } else { NAME <- "BM" } }
+  else if(length(tau)==0)
   { NAME <- "IID" }
 
   # isotropy
@@ -338,6 +392,7 @@ sort.ctmm <- function(x,decreasing=FALSE,IC="AICc",MSPE="position",flatten=TRUE,
 ############
 min.ctmm <- function(x,IC="AICc",MSPE="position",...)
 { sort.ctmm(x,IC=IC,MSPE=MSPE,...)[[1]] }
+
 
 ########
 summary.ctmm.list <- function(object, IC="AICc", MSPE="position", units=TRUE, ...)